# openapi/openapi.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "PagoPA API Spontaneous Payment",
		"description": "Progetto Gestione Pagamenti Spontanei",
		"termsOfService": "https://www.pagopa.gov.it/",
		"version": "0.0.1"
	},
	"servers": [
		{
			"url": "http://localhost:9090",
			"description": "Generated server url"
		}
	],
	"paths": {
...
		"/services": {
			"get": {
				"tags": [
					"Services API"
				],
				"summary": "Return all services.",
				"operationId": "getServices",
				"responses": {
					"200": {
						"description": "Obtained all enrollments for the creditor institution.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrganizationModelResponse"
								}
							}
						}
					},
					"500": {
						"description": "Service unavailable.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemJson"
								}
							}
						}
					},
					"401": {
						"description": "Wrong or missing function key.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [
					{
						"ApiKey": []
					},
					{
						"Authorization": []
					}
				]
			},
			"parameters": [
				{
					"name": "X-Request-Id",
					"in": "header",
					"description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
					"schema": {
						"type": "string"
					}
				}
			]
		},
		"/services/{serviceId}": {
			"get": {
				"tags": [
					"Services API"
				],
				"summary": "Return the single service details.",
				"operationId": "getServiceDetails",
				"parameters": [
					{
						"name": "serviceId",
						"in": "path",
						"description": "The service id for which to have the details.",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Obtained single service details.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceDetailModelResponse"
								}
							}
						}
					},
					"404": {
						"description": "No service found.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemJson"
								}
							}
						}
					},
					"500": {
						"description": "Service unavailable.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemJson"
								}
							}
						}
					},
					"401": {
						"description": "Wrong or missing function key.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [
					{
						"ApiKey": []
					},
					{
						"Authorization": []
					}
				]
			},
			"parameters": [
				{
					"name": "X-Request-Id",
					"in": "header",
					"description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
					"schema": {
						"type": "string"
					}
				}
			]
		}
	},
	"components": {
		"schemas": {
...
			"ProblemJson": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string",
						"description": "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable"
					},
					"status": {
						"maximum": 600,
						"minimum": 100,
						"type": "integer",
						"description": "The HTTP status code generated by the origin server for this occurrence of the problem.",
						"format": "int32",
						"example": 200
					},
					"detail": {
						"type": "string",
						"description": "A human readable explanation specific to this occurrence of the problem.",
						"example": "There was an error processing the request"
					}
				}
			},
			"EnrollmentModelResponse": {
				"required": [
					"iban",
					"serviceId"
				],
				"type": "object",
				"properties": {
					"serviceId": {
						"type": "string"
					},
					"iban": {
						"type": "string"
					},
					"officeName": {
						"type": "string"
					}
				}
			},
			"OrganizationModelResponse": {
				"required": [
					"companyName",
					"fiscalCode",
					"status"
				],
				"type": "object",
				"properties": {
					"fiscalCode": {
						"type": "string"
					},
					"companyName": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"ENABLED",
							"DISABLED"
						]
					},
					"enrollments": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/EnrollmentModelResponse"
						}
					}
				}
			},
...
			"ServiceDetailModelResponse": {
				"required": [
					"id",
					"remittanceInformation",
					"status",
					"transferCategory"
				],
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"transferCategory": {
						"type": "string"
					},
					"remittanceInformation": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"ENABLED",
							"DISABLED"
						]
					},
					"properties": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ServicePropertyModelResponse"
						}
					}
				}
			},
			"ServicePropertyModelResponse": {
				"required": [
					"name"
				],
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"required": {
						"type": "boolean"
					}
				}
			}
		},
		"securitySchemes": {
			"ApiKey": {
				"type": "apiKey",
				"description": "The API key to access this function app.",
				"name": "Ocp-Apim-Subscription-Key",
				"in": "header"
			},
			"Authorization": {
				"type": "http",
				"description": "JWT token get after Azure Login",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}